GFDNet/requirements.txt
timm
catalyst==20.09
pytorch-lightning==1.5.9
albumentations==1.1.0
ttach
numpy
tqdm
opencv-python
scipy
matplotlib
einops
addict
thop
